# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nova Hub",
...
    "contact": {
      "name": "Nova Hub",
      "url": "https://github.com/yourusername/nova-hub"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Access Token",
        "description": "OAuth2 token endpoint for client authentication\n\n**Grant Type:** client_credentials\n\n**Form Parameters:**\n- `grant_type`: Must be \"client_credentials\"\n- `client_id`: Your client ID\n- `client_secret`: Your client secret\n\n**Returns:** Access token valid for 24 hours\n\n**Example:**\n```bash\ncurl -X POST \"https://hub.example.com/auth/token\" \\\n  -d \"grant_type=client_credentials\" \\\n  -d \"client_id=your_client_id\" \\\n  -d \"client_secret=your_secret\"\n```",
        "operationId": "login_for_access_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify Access Token",
        "description": "Verify your access token and get client information\n\n**Requires:** Valid Bearer token in Authorization header\n\n**Returns:** Client details if token is valid\n\n**Example:**\n```bash\ncurl -X GET \"https://hub.example.com/auth/verify\" \\\n  -H \"Authorization: Bearer YOUR_ACCESS_TOKEN\"\n```",
        "operationId": "verify_token_auth_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_for_access_token_auth_token_post": {
        "properties": {
          "grant_type": {
            "type": "string",
            "title": "Grant Type",
            "description": "Must be 'client_credentials'"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "Your client ID provided by the hub administrator"
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret",
            "description": "Your client secret provided by the hub administrator"
          }
        },
        "type": "object",
        "required": [
          "grant_type",
          "client_id",
          "client_secret"
        ],
        "title": "Body_login_for_access_token_auth_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}